Project Final
Here is my project Final, I was not able to test it on a phone and
couldnt get the simulator working. I also still have a little work to do
on it, I have it down to about 7 problems when i build but i think i
willl have more when I fix those but they are just small tweaks that i
will be able to fix by reviewing the code and tutorials. I didn't have
time to make a video but I also didn't want to publish it to Youtube
because I still plan on working on it more and publishing it. I also
have to fix the User Interface and make it look better. The parse
database doesn't have any data but the code should be there. I decided
to go with both a  Parse database and a local Sql because I didn't feel
comfortable with my ability to add data to Parse from a running
application.